♻️ [#32] Refactor: Input 컴포넌트를 역할별로 파일 분리
